+++ packages/isaacscript-common/src/enums/MysteriousPaperEffect.ts
@@ -1,0 +1,12 @@
+/**
+ * The possible effects that the Mysterious Paper trinket can grant.
+ *
+ * This enum has hard-coded values because they correspond to the specific in-game frame count of
+ * the player.
+ */
+export enum MysteriousPaperEffect {
+  POLAROID = 0,
+  NEGATIVE = 1,
+  MISSING_PAGE = 2,
+  MISSING_POSTER = 3,
+}