+++ packages/isaacscript-common/src/functions/revive.ts
@@ -10,11 +10,13 @@ import {
   MAX_TAINTED_SAMSON_BERSERK_CHARGE,
   TAINTED_SAMSON_BERSERK_CHARGE_FROM_TAKING_DAMAGE,
 } from "../core/constants";
+import { MysteriousPaperEffect } from "../enums/MysteriousPaperEffect";
 import { getCharacterDeathAnimationName } from "./characters";
 import { getPlayerMaxHeartContainers } from "./playerHealth";
 import { getPlayerNumHitsRemaining, hasLostCurse, isKeeper } from "./players";
 import { getLastFrameOfAnimation } from "./sprites";
 import { giveTrinketsBack, temporarilyRemoveTrinket } from "./trinketGive";
+import { getMysteriousPaperEffectForFrame } from "./trinkets";
 
 /**
  * Uses the player's current health and other miscellaneous things to determine if incoming damage
@@ -123,7 +125,6 @@ export function isDamageToPlayerFatal(
  * Mysterious Paper be Missing Power is: `gameFrameCount % 4 === 3`
  */
 export function willMysteriousPaperRevive(player: EntityPlayer): boolean {
-  const gameFrameCount = game.GetFrameCount();
   const sprite = player.GetSprite();
 
   // We want to explicitly check the length of the death animation because we might be playing on a
@@ -131,10 +132,17 @@ export function willMysteriousPaperRevive(player: EntityPlayer): boolean {
   const character = player.GetPlayerType();
   const animation = getCharacterDeathAnimationName(character);
   const deathAnimationFrames = getLastFrameOfAnimation(sprite, animation);
-  const frameOfDeath = gameFrameCount + deathAnimationFrames + 1;
-  // (We add 1 because it takes one frame for the death animation to begin.)
+  const frameOfDeath = player.FrameCount + deathAnimationFrames;
 
-  return frameOfDeath % 4 === 3;
+  const mysteriousPaperEffect = getMysteriousPaperEffectForFrame(
+    player,
+    frameOfDeath,
+  );
+  if (mysteriousPaperEffect === undefined) {
+    return false;
+  }
+
+  return mysteriousPaperEffect === MysteriousPaperEffect.MISSING_POSTER;
 }
 
 /**