+++ packages/isaacscript-common/src/index.ts
@@ -11,6 +11,7 @@ export * from "./enums/HealthType";
 export * from "./enums/ISCFeature";
 export * from "./enums/LadderSubTypeCustom";
 export * from "./enums/ModCallbackCustom";
+export * from "./enums/MysteriousPaperEffect";
 export * from "./enums/PocketItemType";
 export * from "./enums/RockAltType";
 export * from "./enums/SaveDataKey";